requirements.txt
# OpenPartSelector Dependencies
# AI & LLM
openai>=1.0.0
anthropic>=0.3.0

# Async HTTP
aiohttp>=3.9.0
httpx>=0.25.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# PDF Parsing
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Vector Database (可选，支持扩展)
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# CLI
click>=8.0.0
rich>=13.0.0

# Logging & Monitoring
loguru>=0.7.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
